Use trgVocab directly
Previously we created our own vocab and then took the intersection of that and trgVocab to construct the trie from. Presumably because then you did not have to preprocess the to-be-aligned document with the same steps (i.e. bpe) as the source input. But then trgVocab and our own vocab would have very little overlap to begin with, and the trie would be pretty useless. We can change it back when we figure out a meaningful on-the-fly-vocab -> trgVocab translation.